// modules/setup/types/company.interface.ts
export interface Company {
    company_id?: number;
    company_logo?: string;
    company_name?: string;
    company_address?: string;
    contact_land?: string;
    contact_mobile?: string;
    contact_whatsapp?: string;
    contact_fax?: string;
    company_email?: string;
    register_number?: string;
    tax_number?: string;
}